Move connectionErrorHandler to tools-mongodb package
The connectionErrorHandler has been moved from mongodb-mcp-server to
the tools-mongodb package where it can be more appropriately shared.
This centralizes the MongoDB connection error handling logic in the
package that defines the MongoDB-specific error types.

Changes:
- Moved connectionErrorHandler.ts to packages/tools-mongodb/src/
- Updated all imports in mongodb-mcp-server to use @mongodb-js/mcp-tools-mongodb
- Updated test file import
- Removed duplicate file from mongodb-mcp-server/src/common/
- Exported connectionErrorHandler from tools-mongodb index.ts
