--- packages/cli/src/handlers/index.ts
@@ -1,5 +1,0 @@
-export type { CliHandler, HandlerResult } from "./types.js";
-export { HelpHandler, handleHelpRequest } from "./helpHandler.js";
-export { VersionHandler, handleVersionRequest } from "./versionHandler.js";
-export { DryRunHandler, handleDryRun, type ServerCreator } from "./dryRunHandler.js";
-export { SetupHandler, type SetupFunction } from "./setupHandler.js";
