+++ packages/cli/src/server/mcpHttpServerWrapper.ts
@@ -1,50 +1,36 @@
-import { MCPHttpServer } from "@mongodb-js/mcp-http-runners";
+import { MCPHttpServer, type SessionAwareServer } from "@mongodb-js/mcp-http-runners";
 import type { SessionStore } from "@mongodb-js/mcp-core";
 import type { HttpServerOptions, SessionManagementOptions } from "@mongodb-js/mcp-types";
 import type { CompositeLogger } from "@mongodb-js/mcp-core";
 import type { IMetrics, DefaultMetricDefinitions } from "@mongodb-js/mcp-types";
-import type { UserConfig } from "../config/userConfig.js";
-import type { ServerCreator } from "../handlers/dryRunHandler.js";
+import type { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
+import type { StartableServer } from "../types.js";
 
-type ServerType = {
-    connect(transport: any): Promise<void>;
-    close(): Promise<void>;
-};
-
-export class MCPHttpServerWrapper extends MCPHttpServer<ServerType> {
-    private userConfig: UserConfig;
-    private baseLogger: CompositeLogger;
-    private createServer: ServerCreator;
+export class MCPHttpServerWrapper extends MCPHttpServer<SessionAwareServer> {
+    private server: SessionAwareServer;
 
     constructor({
-        userConfig,
-        createServer,
+        server,
         options,
         logger,
         metrics,
         sessionStore,
     }: {
-        userConfig: UserConfig;
-        createServer: ServerCreator;
+        server: StartableServer;
         options: {
             http: HttpServerOptions;
             session: SessionManagementOptions;
         };
         logger: CompositeLogger;
         metrics: IMetrics<DefaultMetricDefinitions>;
-        sessionStore: SessionStore<any>;
+        sessionStore: SessionStore<StreamableHTTPServerTransport>;
     }) {
         super({ options, logger, metrics, sessionStore });
-        this.userConfig = userConfig;
-        this.baseLogger = logger;
-        this.createServer = createServer;
+        // Cast to SessionAwareServer since we know the server satisfies the interface at runtime
+        this.server = server as SessionAwareServer;
     }
 
-    protected override async createServerForRequest(): Promise<ServerType> {
-        return this.createServer({
-            config: this.userConfig,
-            logger: this.baseLogger,
-            metrics: this.metrics,
-        });
+    protected override async createServerForRequest(): Promise<SessionAwareServer> {
+        return this.server;
     }
 }
