+++ packages/mongodb-mcp-server/src/index.ts
@@ -34,16 +34,16 @@ function enableFipsIfRequested(): void {
     }
 }
 
+enableFipsIfRequested();
+
 import crypto from "crypto";
 import { ConsoleLogger } from "@mongodb-js/mcp-logging";
-import { LogId } from "@mongodb-js/mcp-core";
-import { Keychain } from "@mongodb-js/mcp-core";
-import { setupMcpCli } from "@mongodb-js/mcp-cli";
+import { LogId, Keychain, Elicitation } from "@mongodb-js/mcp-core";
+import { runMcpCli, createServerFromUserConfig, type Handler, type UserConfig } from "@mongodb-js/mcp-cli";
 import { Server } from "./server.js";
 import { Session } from "./common/session.js";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { Elicitation } from "@mongodb-js/mcp-core";
-import { connectionErrorHandler } from "./common/connectionErrorHandler.js";
+import { connectionErrorHandler } from "@mongodb-js/mcp-tools-mongodb";
 import { MCPConnectionManager, ExportsManager } from "@mongodb-js/mcp-tools-mongodb";
 import { createAtlasLocalClient } from "@mongodb-js/mcp-tools-atlas-local";
 import { ApiClient } from "@mongodb-js/mcp-atlas-api-client";
@@ -53,92 +53,38 @@ import { DeviceId } from "@mongodb-js/mcp-tools-mongodb";
 import { runSetup } from "./setup/setupMcpServer.js";
 import { packageInfo } from "./common/packageInfo.js";
 
-const createServer = async ({ config, logger, metrics }: { config: any; logger: any; metrics: any }) => {
-    const keychain = CoreKeychain.root;
-
-    const exportsManager = ExportsManager.init({
-        options: {
-            exportsPath: config.exportsPath,
-            exportTimeoutMs: config.exportTimeoutMs,
-            exportCleanupIntervalMs: config.exportCleanupIntervalMs,
-        },
-        logger,
-    });
-
-    const deviceId = DeviceId.create(logger);
-
-    const connectionManager = new MCPConnectionManager({
-        logger,
-        deviceId,
-        options: {
-            connectionInfo: { transport: "http", httpHost: "localhost" },
-            displayName: "mongodb-mcp-server",
-            version: packageInfo.version,
-        },
-    });
-
-    const apiClient = new ApiClient({
-        baseUrl: config.apiBaseUrl,
-        userAgent: `mongodb-mcp-server/${packageInfo.version}`,
-        logger,
-        credentials: {
-            clientId: config.apiClientId,
-            clientSecret: config.apiClientSecret,
-        },
-    });
-
-    const atlasLocalClient = await createAtlasLocalClient({
-        logger,
-    });
-    const telemetry = AtlasTelemetry.create({
-        logger,
-        deviceId,
-        apiClient,
-        keychain,
-        enabled: config.telemetry === "enabled",
-        machineMetadata: buildMachineMetadata(packageInfo.mcpServerName, packageInfo.version),
-    });
-
-    const mcpServer = new McpServer({
-        name: "mongodb-mcp-server",
-        version: packageInfo.version,
-    });
-
-    const elicitation = new Elicitation({ server: mcpServer.server });
-
-    const session = new Session({
-        userConfig: config,
-        logger,
-        exportsManager,
-        connectionManager,
-        keychain,
-        apiClient,
-        connectionErrorHandler,
-        atlasLocalClient,
-    });
-
-    const server = new Server({
-        session,
-        userConfig: config,
-        mcpServer,
-        telemetry,
-        connectionErrorHandler,
-        elicitation,
-        metrics,
-    });
-
-    return server;
+const setupHandler: Handler = {
+    shouldHandle(_config: UserConfig, args: string[]): boolean {
+        return args[0] === "setup";
+    },
+    async handle(config: UserConfig, consoleLogger, onExit): Promise<void> {
+        await runSetup(config);
+        onExit(0);
+    },
 };
 
-setupMcpCli({
-    packageInfo: {
-        version: packageInfo.version,
-        mcpServerName: packageInfo.mcpServerName,
-    },
-    createServer,
-    setupFunction: runSetup,
-    enableFipsIfRequested,
-}).catch((error: unknown) => {
+async function main(): Promise<void> {
+    const args = process.argv.slice(2);
+
+    const { config, logger, metrics } = await createServerFromUserConfig({ args, consoleLogger: console });
+
+    await runMcpCli({
+        args,
+        consoleLogger: console,
+        onExit: (code) => process.exit(code),
+        clientInfo: {
+            name: packageInfo.mcpServerName,
+            version: packageInfo.version,
+        },
+        handlers: [setupHandler],
+        server,
+        config,
+        logger,
+        metrics,
+    });
+}
+
+main().catch((error: unknown) => {
     const logger = new ConsoleLogger({ keychain: Keychain.root });
     logger.emergency({
         id: LogId.serverStartFailure,
